Move and recategorize tests
