+++ .github/workflows/main.yml
@@ -455,7 +455,7 @@ jobs:
           sudo rm -rf "/usr/local/share/boost" "$AGENT_TOOLSDIRECTORY" "/usr/local/lib/android" "/usr/share/dotnet"
           docker run --rm -i --name libvcx --network host $DOCKER_IMG_NAME_LIBVCX \
                               bash -c '(cd $HOME/libvcx && \
-                                RUST_TEST_THREADS=1 cargo test --release --features "general_test aries" && \
+                                RUST_TEST_THREADS=1 cargo test --release --features "general_test" && \
                                 cd $HOME/agency_client && \
                                 RUST_TEST_THREADS=1 cargo test --release --features "general_test")'
