+++ libvcx/Cargo.toml
@@ -20,7 +20,6 @@ agency = [] # migrate these to v2 agency
 agency_pool_tests = []
 agency_v2 = []
 pool_legacy_agency_tests = [] # migrate the test using this
-aries = []
 general_test = []
 to_restore = []
 # Causes the build to fail on all warnings
